pyproject.toml
[tool.poetry]
name = "visualization-service"
version = "0.1.0"
description = ""
authors = ["Timon <timonbub@gmail.com>"]

packages = [
    { include = "visualization_service", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"

# data
pandas = "^2.3"
matplotlib = "^3.10"
seaborn = "^0.13"
psycopg2-binary = "^2.9"

# API
fastapi = "^0.110"
uvicorn = { extras = ["standard"], version = "^0.27" }

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
